+++ json_serializable/lib/src/generator_helper.dart
@@ -301,7 +301,8 @@ void $toJsonMapHelperName(String key, dynamic value) {
   }
 
   TypeHelperContext _getHelperContext(FieldElement field) =>
-      new TypeHelperContext(_generator, field.metadata, _nullable(field));
+      new TypeHelperContext(
+          _generator, field.metadata, _nullable(field), jsonKeyFor(field));
 
   /// Returns `true` if the field can be written to JSON 'naively' – meaning
   /// we can avoid checking for `null`.