+++ json_serializable/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.5.1
 
+* Support new `fromJson` and `toJson` fields on `JsonKey`.
+
 * Use `log` exposed by `package:build`. This requires end-users to have at least
   `package:build_runner` `^0.8.2`.
+
 * Updated minimum `package:source_gen` dependency to `0.8.1` which includes
   improved error messages.